requirements.txt
dj-database-url==0.5.0
Django==2.1.5
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2
django-froala-editor==3.1.0
django-richtextfield==1.5.0
django-summernote==0.8.11.6
django-tinymce==3.0.2
django-embed-video==1.3.2
Pillow==7.1.2